Add artificial intelligence capabilities to make important life choices
By using state of the art natural language processing techniques that's
able to understand both Portuguese, English, and a subset of the C
programming language, this new feature will enable the bot to make
important life decisions.

To use it, just ask the bot:

	<you> nickname, do this or do that?
	<nickname> you: do this!

It's not advisable to ask the bot again, as the slightest change in the
state of the universe might alter the calculations and yield a
different response.